JSS-92 Display the commit hash and commit link on the page
We now display the latest commit hash which is linked to the latest
commit on the master branch of JSS.

We use the latest commit hash as that is the commit of the latest NuGet
package (which is automatically uploading to NuGet). However, we do have
to manually merge the dependabot PRs for the JSS package updates.

We should try to automatically merge the PRs from dependabot for JSS
package updates.